app/requirements_multi_agent.txt
fastapi
langchain>=0.1.16
langchain-openai>=0.1.3
langgraph
pydantic
pytest
python-dotenv>=1.0
streamlit>=1.33
uvicorn[standard]
